Move ValidatedNotNullAttribute from snippet testing to test file.
